refactor: 이미지 슬라이더 모듈화
